+++ taz.neo/Helper/PushNotification.Payload+Extension.swift
@@ -52,7 +52,7 @@ extension PushNotification.Payload {
     }
   }
   
-  typealias ArticlePushData = (articleMsId:Int,
+  typealias ArticlePushData = (articleMsId:Int64,
                                articleTitle:String?,
                                articleBody:String?,
                                articleDate:Date,
@@ -69,7 +69,7 @@ extension PushNotification.Payload {
 extension [AnyHashable:Any] {
   var articlePushData: PushNotification.Payload.ArticlePushData? {
     get {
-      guard let articleMsId = self["articleMsId"] as? Int else { return nil }
+      guard let articleMsId = self["articleMsId"] as? Int64 else { return nil }
       let articleTitle = self["articleTitle"] as? String
       let articleBody = self["articleBody"] as? String
       guard let articleDateString = self["articleDate"] as? String else { return nil }
